requirements.txt
streamlit
pandas
numpy
yfinance
plotly
google-generativeai
python-dotenv
GoogleNews
newsapi-python
lxml